refactor(Auth, UI): UserProfile 컴포넌트 분리 및 이니셜 생성 로직 개선
UserProfile 컴포넌트를 Header.tsx에서 src/components/UserProfile.tsx 파일로 분리하여 모듈성을 향상시키고 Header 컴포넌트의 가독성을 개선했습니다.

또한, UserProfile 컴포넌트 내에서 사용자 이메일로 프로필 이미지 이니셜을 생성하는 로직을 수정했습니다. 기존에는 이메일 길이가 1인 경우 `substring(0, 2)` 호출 시 오류가 발생할 수 있었으나, 이메일 길이를 확인하여 안전하게 이니셜을 생성하도록 변경했습니다.

주요 변경 사항:
- UserProfile 컴포넌트를 `src/components/UserProfile.tsx`로 이동
- UserProfile 컴포넌트의 `getInitials` 함수에서 이메일 길이가 1일 때 `substring` 오류 방지 로직 추가

PR 리뷰 반영 #10
